🟢 AlertDetails.jsx 🟢 AlertDetails.module.css 🟢 index.js 🛠️ AlertLog.jsx -> Removed AlertDetails code 🛠️ AlertLog.jsx -> Imported AlertDetails 🛠️ AlertLog.module.css -> Removed styles
