+++ src/app/components/AlertDetails/index.js
@@ -1,0 +1,1 @@
+export * from './AlertDetails';
