+++ src/app/components/AlertLog/AlertLog.jsx
@@ -1,9 +1,7 @@
 "use client";
 
-import React, { useState } from "react";
+import React from "react";
 import {
-  RuxButton,
-  RuxDialog,
   RuxTable,
   RuxTableBody,
   RuxTableCell,
@@ -12,56 +10,11 @@ import {
   RuxTableHeaderRow,
   RuxTableRow,
 } from "@astrouxds/react";
+import { AlertDetails } from "../AlertDetails";
 
 import styles from "./AlertLog.module.css";
 
-const AlertDetails = ({ alert, contact, doUpdateEntry }) => {
-  const [isAcknowledged, setIsAcknowledged] = useState(false);
-
-  const buttonText = alert.acknowledged
-    ? "Review Details"
-    : "Acknowledge Alert";
-  const dialogButtonText = alert.acknowledged ? "Dismiss" : "Acknowledge";
-
-  const dialogClosed = () => {
-    setIsAcknowledged(false);
-    doUpdateEntry({ alert, contact });
-  };
-
-  return (
-    <>
-      <RuxButton
-        size="small"
-        secondary={alert.acknowledged}
-        borderless={alert.acknowledged}
-        onClick={() => setIsAcknowledged(true)}
-      >
-        {buttonText}
-      </RuxButton>
-
-      <RuxDialog
-        open={isAcknowledged}
-        header={alert.errorMessage}
-        confirmText={dialogButtonText}
-        denyText=""
-        onRuxdialogclosed={dialogClosed}
-      >
-        <div className={styles.AlertDetailText}>
-          <h3>Contact Satellite: {contact.contactSatellite}</h3>
-          <p>{contact.contactDetail}</p>
-          <p className={styles.ErrorTime}>
-            <label className={styles.Label}>Error Time:</label>{" "}
-            {alert.errorTimeString}
-          </p>
-        </div>
-      </RuxDialog>
-    </>
-  );
-};
-
 const AlertLog = ({ data = [], doUpdateEntry }) => {
-
-
 
   // Returns CSS class based on the severity level of an alert
   const severityClasses = (severityLevel) => {
